+++ CHANGELOG.md
@@ -4,6 +4,13 @@ TODO
 - type "becomes" to the property value
 - maybe a way to type rule objects without starting (like itemEffect() as an import?)
 
+v0.9.2
+
+- added helpers to make easier switch-like rules with makeRuleMaker, makeNestedRuleMaker, makeNestedLeaveRuleMaker NOTE if it's too much, they could be made with a seperate rules helper that takes setState and makeRule etc as parameters
+- now state properties can be typed based on the item name (from startStates), like a model.animationWeights object can have the correct animation names as keys for that object, before it would combine all the animation names from all the models
+- added "run" to rules objects to run a rule as a manual action, good as a start to being able to run all rules on initial state :)
+- fixed rule names for rules objects not being typed
+
 v0.9.1
 
 - renamed to repond and updated names to match